Refactor: JVM 관리 로직 infra 디렉으로 이동해라 - #94
