--- src/main/java/talkPick/infra/cache/model/topic/Age.java
@@ -1,13 +1,0 @@
-package talkPick.infra.cache.model.topic;
-
-public enum Age {
-    TEENS,      // 10대
-    TWENTIES,   // 20대
-    THIRTIES,   // 30대
-    FORTIES,    // 40대
-    FIFTIES,    // 50대
-    SIXTIES,    // 60대
-    SEVENTIES,  // 70대
-    EIGHTIES,   // 80대
-    NINETIES;   // 90대
-}
